fix: errors BJS2-65898
