+++ src/main/java/school/faang/user_service/mapper/education/EducationMapper.java
@@ -1,10 +1,11 @@
 package school.faang.user_service.mapper.education;
 
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 import school.faang.user_service.dto.education.EducationDto;
 import school.faang.user_service.entity.Education;
 
-@Mapper(componentModel = "spring")
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface EducationMapper {
 
     Education toEducation(EducationDto educationDto);
