--- src/main/java/school/faang/user_service/service/user/UserService.java
@@ -1,22 +1,0 @@
-package school.faang.user_service.service.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.exception.DataValidationException;
-import school.faang.user_service.repository.UserRepository;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserRepository userRepository;
-
-    public User findById(long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new DataValidationException("User not found"));
-    }
-
-    public boolean existsById(long userId) {
-        return userRepository.existsById(userId);
-    }
-}
