+++ src/test/java/school/faang/user_service/service/education/EducationServiceTest.java
@@ -11,7 +11,7 @@ import school.faang.user_service.entity.User;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.mapper.education.EducationMapper;
 import school.faang.user_service.repository.EducationRepository;
-import school.faang.user_service.service.user.UserService;
+import school.faang.user_service.service.UserService;
 
 import java.time.Year;
 import java.util.Optional;
